Diagrams.Coordinates.Isomorphic

Description

This module defines a class for coordinates that are (loosely) isomorphic to the standard spaces (V2 and V3). This allows plots to accept more data types for plot data.

Type constraints

type HasIndexedBasis v = (HasBasis v, TraversableWithIndex (E v) v)

type Euclidean v = (HasLinearMap v, HasIndexedBasis v, Metric v)

Umbrella class giving everything needed for working in the space. This is basically V2 or V3 from "linear".

Vector like

class (Euclidean v, Typeable v) => VectorLike v n a | a -> v n where

Provides an Iso' between a and v n. This is normally used to convert between the data type you're already using, a, and diagram's native form, v n.

Minimal complete definition

vectorLike

Methods

vectorLike :: Iso' (v n) a

Isomorphism from Point v n to something PointLike a.

>>> V2 3 5 ^. vectorLike :: (Int, Int)
(3,5)

unvectorLike :: Iso' a (v n)

Isomorphism from something PointLike a to Point v n.

>>> ((3, 5) :: (Int, Int)) ^. unvectorLike
V2 3 5

type V2Like = VectorLike V2

type V3Like = VectorLike V3

Point like

class (Euclidean v, Typeable v) => PointLike v n a | a -> v n where

Provides an Iso' between a and Point v n. This is normally used to convert between the data type you're already using, a, and diagram's native form, Point v n.

Minimal complete definition

pointLike

Methods

pointLike :: Iso' (Point v n) a

Isomorphism from Point v n to something PointLike a.

>>> mkP2 3 5 ^. pointLike :: (Int, Int)
(3,5)

unpointLike :: Iso' a (Point v n)

Isomorphism from something PointLike a to Point v n.

>>> ((3, 5) :: (Int, Int)) ^. unpointLike
P (V2 3 5)

type P2Like = PointLike V2

Things that are isomorphic to points in R2.

type P3Like = PointLike V3