Diagrams.Coordinates.Polar

Description

This module defines a polar coordinate data type. This type can be used as an axis space for polar plots.

Polar type

newtype Polar a

Constructors

Polar (V2 a)

mkPolar :: n -> Angle n -> Polar n

Construct a Polar from a magnitude and an Angle.

polar :: (n, Angle n) -> Polar n

Construct a Polar from a magnitude and Angle tuple.

unpolar :: Polar n -> (n, Angle n)

Turn a Polar back into a magnitude and Angle tuple.

polarIso :: Iso' (Polar n) (n, Angle n)

Iso' between Polar and its tuple form.

polarV2 :: RealFloat n => Iso' (Polar n) (V2 n)

Numerical Iso' between Polar and R2.

Polar functions

interpPolar :: Num n => n -> Polar n -> Polar n -> Polar n

Polar interpolation between two polar coordinates.

Classes

class Radial t where

Space which has a radial length basis. For Polar and Cylindrical this is the radius of the circle in the xy-plane. For Spherical this is the distance from the origin.

Methods

_radial :: Lens' (t a) a

class Radial t => Circle t where

Space which has a radial and angular basis.

Methods

_azimuth :: Lens' (t a) (Angle a)

_polar :: Lens' (t a) (Polar a)

class HasX t where

Coordinate with at least one dimension where the x coordinate can be retreived numerically. Note this differs slightly from R1 which requires a lens for all values. This allows instances for different coordinates such as Polar, where the x coordinate can only be retreived numerically.

Minimal complete definition

x_

Methods

x_ :: RealFloat n => Lens' (t n) n

class HasX t => HasY t where

Coordinate with at least two dimensions where the x and y coordinates can be retreived numerically.

Minimal complete definition

xy_

Methods

y_ :: RealFloat n => Lens' (t n) n

xy_ :: RealFloat n => Lens' (t n) (V2 n)

class HasR t where

A space which has magnitude _r that can be calculated numerically.

Minimal complete definition

Nothing

Basis elements

er :: Radial v => E v

eθ :: Circle v => E v

etheta :: Circle v => E v