Plots

Description

This module defines types for axis labels and tick labels.

Documentation

Axis definition (r2Axis and polarAxis), aspect ratio and scaling.

module Plots.Axis

AxisStyles are used to provide default colours and shapes for the plots of an axis.

module Plots.Style

Plot Types

Scatter plot

Scatter plots display data as a collection of points. A scatter plot can also be configured to have a different style / transform depending on the data.

Scatter and bubble. Scatter and bubble plot api.

module Plots.Types.Scatter

Bar plots, individual or grouped.

module Plots.Types.Bar

Line plot

Line, trail and path. Line plot, steps plot api & api for trail and path.

module Plots.Types.Line

Heat map plot

2D mapping from Doubles to colours.

module Plots.Types.HeatMap

Histogram plot

Histogram. API for histogram.

module Plots.Types.Histogram

Wedge and annular wedge. API for wedge, annular wegde and pie.

module Plots.Types.Pie

Low level

Definitions of bounds, axis scale, orientation, legend, generic plot ,plot spec and so on.

module Plots.Types

Grid lines and styles.

module Plots.Legend

Grid lines and styles.

module Plots.Axis.Grid

Axis labels and tick labels .

module Plots.Axis.Labels

Rendering system for polar and r2 axis.

module Plots.Axis.Render

The scaling/size options for an axis.

module Plots.Axis.Scale

Ticks properties and placement.

module Plots.Axis.Ticks

The plot title.

module Plots.Axis.Title

Colour bars.

module Plots.Axis.ColourBar

Polar coordinates

module Diagrams.Coordinates.Polar

(&=) :: MonadState s m => ASetter' s b -> State b a -> m () infix 3

Similar to '(%=)' but takes a state modification instead of a function.

(&~~) :: Monad m => s -> StateT s m a -> m s infix 1

Similar to '(&~)' but works with StateT and returns it in m.