Plots.Axis

Description

The Axis is the main data type for "plots". It holds all the nessesary infomation to be rendered into a Diagram.

Axis type

data Axis b c n

Axis is the data type that holds all the nessessary information to render a plot. Common LensLikes used for the axis (see haddock's instances for a more comprehensive list):

axisStyle - customise the AxisStyle

legend - customise the Legend

colourBar - customise the ColourBar

currentPlots - current plots in the Axis

finalPlots - changes to the plots just before rendering

axes - changes to each SingleAxis

xAxis - the x-axis

yAxis - the y-axis

zAxis - the z-axis

The following LensLikes can be used on the on all the axes by applying it the to Axis or can be used on a SingleAxis by using it in combination with a specific axis (like xAxis).

axisLabel - customise the MinorTicks

tickLabel - customise the TickLabels

minorTicks - customise the MinorTicks

majorTicks - customise the MajorTicks

gridLines - customise the GridLines

axisLine - customise the AxisLine

axisScaling - customise the AxisScaling

axes :: (v ~ BaseSpace c, v ~ BaseSpace c') => Lens (Axis b c n) (Axis b c' n) (c (SingleAxis b v n)) (c' (SingleAxis b v n))

Lens onto the separate axes of an axis. Allows changing the coordinate system as long as the BaseSpace is the same.

axes :: Lens' (Axis b c n) (c (SingleAxis b v n))

axisPlots :: (BaseSpace c ~ v) => Lens' (Axis b c n) [DynamicPlot b v n]

The list of plots currently in the axis.

currentPlots :: (BaseSpace c ~ v) => Traversal' (Axis b c n) (DynamicPlot b v n)

Traversal over the current plots in the axis.

finalPlots :: (BaseSpace c ~ v) => Setter' (Axis b c n) (DynamicPlot b v n)

Setter over the final plot before the axis is rendered.

For example, to make all ScatterPlots in the axis use a connectingLine, you can add

finalPlots . connectingLine .= True

at the begining of the axis and all scatter plots added will have a connectingLine.

axisSize :: (HasLinearMap c, Num n, Ord n) => Lens' (Axis b c n) (SizeSpec c n)

The size used for the rendered axis.

Predefined axes

r2Axis :: (TypeableFloat n, Enum n, Renderable (Text n) b, Renderable (Path V2 n) b) => Axis b V2 n

The default axis for plots in the V2 coordinate system.

Base space

type family BaseSpace c :: * -> *

This family is used so that we can say (Axis Polar) but use V2 for the underlying diagram.

Axis plots

addPlot

Arguments

:: (InSpace (BaseSpace c) n p, MonadState (Axis b c n) m, Plotable p b)

=> Plot p b
the plot

-> m ()
add plot to the Axis

Add a Plotable Plot to an Axis.

addPlotable

Arguments

:: (InSpace (BaseSpace c) n p, MonadState (Axis b c n) m, Plotable p b)

=> p
the raw plot

-> State (Plot p b) ()
changes to the plot

-> m ()
add plot to the Axis

Add something Plotable to the Axis with a statefull modification of the Plot.

addPlotable'

Arguments

:: (InSpace (BaseSpace v) n p, MonadState (Axis b v n) m, Plotable p b)

=> p
the raw plot

-> m ()
add plot to the Axis

Simple version of AddPlotable without any changes Plot.

Single axis

data SingleAxis b v n

Render infomation for a single axis line.

Specific axes

x-axis

xAxis :: R1 c => Lens' (Axis b c n) (SingleAxis b (BaseSpace c) n)

Lens onto the x-axis of an Axis.

xAxisLabel :: R1 c => Lens' (Axis b c n) String

The label for the x-axis. Shorthand for xAxis . axisLabelText.

xMin :: R1 c => Lens' (Axis b c n) (Maybe n)

The minimum x value for the axis. If the value if Nothing (the Default), the bounds will be infered by the plots in the axis.

xMax :: R1 c => Lens' (Axis b c n) (Maybe n)

The minimum x value for the axis. If the value if Nothing (the Default), the bounds will be infered by the plots in the axis.

y-axis

yAxis :: R2 c => Lens' (Axis b c n) (SingleAxis b (BaseSpace c) n)

Lens onto the y-axis of an Axis.

yAxisLabel :: R2 c => Lens' (Axis b c n) String

The label for the y-axis. Shorthand for yAxis . axisLabelText.

yMin :: R2 c => Lens' (Axis b c n) (Maybe n)

The minimum y value for the axis. If the value if Nothing (the Default), the bounds will be infered by the plots in the axis.

yMax :: R2 c => Lens' (Axis b c n) (Maybe n)

The minimum y value for the axis. If the value if Nothing (the Default), the bounds will be infered by the plots in the axis.

z-axis

zAxis :: R3 c => Lens' (Axis b c n) (SingleAxis b (BaseSpace c) n)

Lens onto the z-axis of an Axis.

zAxisLabel :: R3 c => Lens' (Axis b c n) String

The label for the z-axis. Shorthand for zAxis . axisLabelText.

zMin :: R3 c => Lens' (Axis b c n) (Maybe n)

The minimum z value for the axis. If the value if Nothing (the Default), the bounds will be infered by the plots in the axis.

zMax :: R3 c => Lens' (Axis b c n) (Maybe n)

The minimum z value for the axis. If the value if Nothing (the Default), the bounds will be infered by the plots in the axis.