Plots.Axis.ColourBar

Description

XXX This module in unfinished and features in it may not work.

Documentation

data ColourBar b n

Options for drawing a colour bar. Note that for an axis, the ColourMap is stored in the AxisStyle. These options are for other aspects of the bar, not the colours used.

Constructors

ColourBar

Fields

cbOrientation :: Orientation

cbVisible :: Bool

cbMajorTicks :: MajorTicks V2 n

cbExtent :: V2 n

cbGap :: n

cbStyle :: Style V2 n

defColourBar :: (Renderable (Text n) b, Renderable (Path V2 n) b, TypeableFloat n, Enum n) => ColourBar b n

class HasColourBar a b | a -> b where

Minimal complete definition

colourBar

Methods

colourBar :: Lens' a (ColourBar b (N a))

Lens onto the ColourBar.

colourBarVisible :: Lens' a Bool

Whether the colour bar should be visible when rendering.

colourBarExtent :: Lens' a (V2 (N a))

The size of the colour bar when it's in a Horizontal orientation.

colourBarGap :: Lens' a (N a)

Gap between the axis and the colour bar (if rendered with an axis).

colourBarStyle :: Lens' a (Style V2 (N a))

Style used for the outline of a colour bar.

drawColourBar

Arguments

:: (TypeableFloat n, Renderable (Path V2 n) b)

=> ColourBar b n
options

-> ColourMap
colours to use

-> (n, n)
bounds for colour bar tick labels

-> QDiagram b V2 n Any
resulting colour bar

Draw a standalone colour bar.

addColourBar :: (TypeableFloat n, Renderable (Path V2 n) b) => BoundingBox V2 n -> ColourBar b n -> ColourMap -> (n, n) -> QDiagram b V2 n Any

Add a colour bar to an object, using the bounding box for the object.