Plots.Axis.Render

Description

Low level module containing functions for rendering different types of axis.

Documentation

buildPlots :: BaseSpace c ~ v => Axis b c n -> [StyledPlot b v n]

Build a list of styled plots from the axis, ready to be rendered. This takes into account any AxisStyle changes and applies the finalPlots modifications.

class RenderAxis b v n where

Minimal complete definition

renderAxis

Methods

renderAxis :: Axis b v n -> QDiagram b (BaseSpace v) n Any

renderR2Axis :: (TypeableFloat n, Renderable (Path V2 n) b) => Axis b V2 n -> QDiagram b V2 n Any

data LabelPosition

The position of axis labels for a

Constructors

NoLabels

LowerLabels

LeftLabels

RightLabels

UpperLabels

axisOnBasis :: forall b v n. (v ~ V2, TypeableFloat n, HasLinearMap v, Metric v, Renderable (Path V2 n) b, n ~ N (v n), v ~ V (v n), OrderedField n) => Point v n -> v (n, n) -> SingleAxis b v n -> v LogScale -> T2 n -> E v -> E v -> LabelPosition -> QDiagram b V2 n Any

primStroke :: (Ord n, Typeable n, Typeable v, Renderable (Path v n) b) => Path v n -> QDiagram b v n Any

Stroke without any envelope, trace, query etc.

translationE :: (Num n, HasLinearMap v) => E v -> n -> Transformation v n

scaleE :: (Additive v, Fractional n) => E v -> n -> Transformation v n

getAxisLinePos :: (Num n, Ord n) => (n, n) -> AxisLineType -> [(AxisPos, n)]

data AxisPos

Constructors

LowerAxis

MiddleAxis

UpperAxis

ep :: E v -> Lens' (Point v x) x

renderPolarAxis :: (TypeableFloat n, Renderable (Path V2 n) b) => Axis b Polar n -> QDiagram b V2 n Any

drawPolarAxis :: forall b n. (Renderable (Path V2 n) b, TypeableFloat n) => AxisSpec V2 n -> Polar (SingleAxis b V2 n) -> QDiagram b V2 n Any

whenever :: Bool -> (a -> a) -> a -> a

Apply a function if the predicate is true.