Plots.Axis.Title

Description

The title used for a plot.

Documentation

data Title b v n

class HasTitle a b | a -> b where

Minimal complete definition

title

Methods

title :: Lens' a (Title b (V a) (N a))

titleText :: Lens' a String

The text used for the title. If the string is empty, no title is drawn.

Default is ""

titleStyle :: Lens' a (Style (V a) (N a))

The style applied to the title.

Default is mempty.

titlePlacement :: Lens' a Placement

The placement of the title against the axis.

Default is mempty.

titleGap :: Lens' a (N a)

The gap between the axis and the title.

Default is mempty.

drawTitle :: TypeableFloat n => BoundingBox V2 n -> Title b V2 n -> QDiagram b V2 n Any

Render the title and place it around the bounding box.