Plots.Types.Bar

Description

A bar plot is a plot that presents data with rectangular bars with lengths proportional to the values that they represent. The bars can be plotted vertically or horizontally.

BarPlot

data BarPlot n

A bar plot for a single set of bars. Multi-bar plots are achieved by having multiple BarPlots. Each bar plot corresponds to a single legend entry. To get multiple bar entries/colours, use multiple BarPlots

barPlot

Arguments

:: (MonadState (Axis b V2 n) m, Plotable (BarPlot n) b, Foldable f)

=> f n
bar heights

-> State (Plot (BarPlot n) b) ()
changes to the bars

-> m ()
changes to the Axis

A add BarPlot to an Axis.

barPlot'

Arguments

:: (MonadState (Axis b V2 n) m, Plotable (BarPlot n) b, Foldable f)

=> f n
bar heights

-> m ()
changes to the Axis

Simple version of barPlot without any modification to the Plot.

namedBarPlot

Arguments

:: (MonadState (Axis b V2 n) m, Plotable (BarPlot n) b, Foldable f)

=> f (String, n)
bar heights with name

-> State (Plot (BarPlot n) b) ()
changes to the bars

-> m ()
changes to the Axis

A add BarPlot to an Axis while naming the bars.

namedBarPlot'

Arguments

:: (MonadState (Axis b V2 n) m, Plotable (BarPlot n) b, Foldable f)

=> f (String, n)
bar heights with name

-> m ()
add plot to the Axis

Simple version of namedBarPlot without any modification to the Plot.

floatingBarPlot

Arguments

:: (MonadState (Axis b V2 n) m, Plotable (BarPlot n) b, Foldable f)

=> f (n, n)
bar limits

-> State (Plot (BarPlot n) b) ()
changes to the bars

-> m ()

Same as barPlot but with lower and upper bounds for the bars.

Bar layout

data BarLayout n

The way an individual bar plot or a group of bars plots are laid out on the axis.

class HasOrientation a => HasBarLayout a where

Class of things that have a modifiable BarLayout.

Minimal complete definition

barLayout

Methods

barLayout :: Lens' a (BarLayout (N a))

Lens onto the BarLayout

barWidth :: Lens' a (N a)

The width bar for single / stacked bars or the width of a group for grouped bar plot.

Default is 0.8

barSpacing :: Lens' a (N a)

The spacing between each bar or group of bars.

Default is 1

barStart :: Lens' a (N a)

The distance from the axis to centre of the first bar.

Default is 1

Multi bars

Adding to axis

multiBars

Arguments

:: (MonadState (Axis b V2 n) m, Plotable (BarPlot n) b, Foldable f, Foldable g)

=> f a
data for multi plot

-> (a -> g n)
extract bar heights from each data set

-> State (MultiBarState b n a) ()
state to make changes to the plot

-> m ()
changes to the Axis

Construct multiple bars, grouped together. See MultiBarState for details on how to customise how the bars are drawn.

data MultiBarState b n a

The MultiBarState is used to set the various options available when building multiple bar plots together. The main functions used to modify this state:

To choose the way the bars are grouped together choose one of

groupedBars - Together in grouped (the default)

stackedBars - On on top of another

stackedEqualBars - stackedBars with the same height

runningBars - each group of bars follows the last

Modify the PlotOptions and PlotStyle of groups of bars with onBars.

Modify the layout of the (groups of) bars with

orientation - Horizontal or vertical bars

barWidth - Width of each (group of) bar(s)

barSpacing - Space between each (group of) bar(s)

barStart - Start of centre of first bar

Add labels to each (group of) bars with labelBars.

Multi bar types

groupedBars :: Fractional n => State (MultiBarState b n a) ()

Bars that are grouped together such that each group is a single barWidth. The bars in a group are touching, see groupedBars' to reduce the width of individual bars.

groupedBars' :: Fractional n => n -> State (MultiBarState b n a) ()

Bars that are grouped together such that each group is a single barWidth. The parameter is the multiplier for the width of individual bars, where groupedBars 1 = groupedBars corresponds to bars in a group touching. reduce the width of individual bars.

stackedBars :: Num n => State (MultiBarState b n a) ()

Bars stacked on top of each other.

stackedEqualBars :: Fractional n => n -> State (MultiBarState b n a) ()

Bars stacked on top of each other where every bar is the given height.

runningBars :: Num n => State (MultiBarState b n a) ()

Normal bars where each data set follows the last.

Modify multi bars

onBars

Arguments

:: (a -> State (PlotMods b V2 n) ())
Modifier the PlotOptions and PlotStyle for the bars associated with the data from a.

-> State (MultiBarState b n a) ()
Changes to each data set when executing multiBars.

Given the data for the bar, modify the properties for the bar that uses that data.

Some common functions to use on the PlotMods:

plotColour - change the colour of the bars

areaStyle - modify the style of the bars

key - add a legend entry for that group of bars

labelBars :: HasLabels a => [String] -> State a ()

Labels to use for each bar (or group of bars) along the axis.

Low level constructors

mkBars :: (Foldable f, Num n) => BarLayout n -> f n -> BarPlot n

Create equidistant bars using the values.

mkFloatingBars :: (Foldable f, Num n) => BarLayout n -> f (n, n) -> BarPlot n

Create equidistant bars with lower and upper bounds for each bar.

mkRunningBars :: Num n => BarLayout n -> [[(n, n)]] -> [BarPlot n]

Create uniform bars from groups of data, placing one group after the other.

mkStackedBars :: Num n => BarLayout n -> [[n]] -> [BarPlot n]

Create uniform bars from groups of data, placing one on top of the other. The first list will be the same as mkUniformBars opts (map (0,) ys), subsequent lists will be placed on top.

mkStackedEqualBars

Arguments

:: Fractional n

=> n
value each bar reaches

-> BarLayout n

-> [[n]]
values

-> [BarPlot n]

Similar to mkMultiStacked but stack has the same height.

mkGroupedBars

Arguments

:: Fractional n

=> n
width factor of individual bars (1 = touching)

-> BarLayout n

-> [[n]]

-> [BarPlot n]

Make bars that are grouped together. Each group of bars is treated as a single bar when using the BarPlotsOpts. There is an addition parameter to adjust the width of each individual bar.