Plots.Types.Function

Function plot options

data FunctionPlotOptions n

functionPlotNumPoints :: HasFunctionPlotOptions c n => Lens' c Int

Parametric plot

data ParametricPlot v n

Constructors

ParametricPlot

Fields

_parametricFunction :: n -> Point v n

_parametricDomain :: (n, n)

_parametricPlotOptions :: FunctionPlotOptions n

mkParametricPlot :: (PointLike v n p, TypeableFloat n) => (n -> p) -> ParametricPlot v n

Create a parametric plot given a function on range (0,5).

parametricDomain :: forall v n. Lens' (ParametricPlot v n) (n, n)

mkParametricRangePlot :: PointLike v n p => (n -> p) -> (n, n) -> ParametricPlot v n

Create a parametric plot given a function and a range.

Create line

createABLine :: Num n => n -> n -> n -> P2 n

Functions to create ab, vertical and horizontal lines.

createHLine :: n -> n -> P2 n

createVLine :: n -> n -> P2 n

Vector plot

data VectorPlot v n

Constructors

VectorPlot

Fields

_vectorV :: v n

_vectorPoint :: (n, n)

_vectorArrows :: ArrowOpts n

mkVectorPlot :: TypeableFloat n => v n -> VectorPlot v n

Plot a given vector at (0,0).

mkVectorPointPlot :: TypeableFloat n => v n -> (n, n) -> VectorPlot v n

Plot a given vector at a given point.

Vector plot lenses

setArrowOpts :: HasVector a v n => Lens' a (ArrowOpts n)

Parametric plot

parametricPlot :: (v ~ BaseSpace c, PointLike v n p, MonadState (Axis b c n) m, Plotable (ParametricPlot v n) b, TypeableFloat n) => (n -> p) -> State (Plot (ParametricPlot v n) b) () -> m ()

parametricRangePlot :: (v ~ BaseSpace c, PointLike v n p, MonadState (Axis b c n) m, Plotable (ParametricPlot v n) b) => (n -> p) -> (n, n) -> State (Plot (ParametricPlot v n) b) () -> m ()

parametricPlot' :: (v ~ BaseSpace c, PointLike v n p, MonadState (Axis b c n) m, Plotable (ParametricPlot v n) b, TypeableFloat n) => (n -> p) -> m ()

parametricRangePlot' :: (v ~ BaseSpace c, PointLike v n p, MonadState (Axis b c n) m, Plotable (ParametricPlot v n) b) => (n -> p) -> (n, n) -> m ()

Line functions

abLinePlot :: (v ~ BaseSpace c, PointLike v n (P2 n), MonadState (Axis b c n) m, Plotable (ParametricPlot v n) b, Additive v, TypeableFloat n) => n -> n -> (n, n) -> State (Plot (ParametricPlot v n) b) () -> m ()

hLinePlot :: (v ~ BaseSpace c, PointLike v n (P2 n), MonadState (Axis b c n) m, Plotable (ParametricPlot v n) b, Additive v, TypeableFloat n) => n -> (n, n) -> State (Plot (ParametricPlot v n) b) () -> m ()

vLinePlot :: (v ~ BaseSpace c, PointLike v n (P2 n), MonadState (Axis b c n) m, Plotable (ParametricPlot v n) b, Additive v, TypeableFloat n) => n -> (n, n) -> State (Plot (ParametricPlot v n) b) () -> m ()

Vectors

vectorPlot :: (v ~ BaseSpace c, MonadState (Axis b c n) m, Plotable (VectorPlot v n) b, TypeableFloat n) => v n -> State (Plot (VectorPlot v n) b) () -> m ()

vectorPointPlot :: (v ~ BaseSpace c, MonadState (Axis b c n) m, Plotable (VectorPlot v n) b, TypeableFloat n) => v n -> (n, n) -> State (Plot (VectorPlot v n) b) () -> m ()

vectorPointPlot' :: (v ~ BaseSpace c, MonadState (Axis b c n) m, Plotable (VectorPlot v n) b, TypeableFloat n) => v n -> (n, n) -> m ()

vectorPointPlot'' :: (v ~ BaseSpace c, MonadState (Axis b c n) m, Plotable (VectorPlot v n) b, TypeableFloat n) => v n -> (n, n) -> ArrowOpts n -> m ()

vectorFieldPlot :: (v ~ BaseSpace c, MonadState (Axis b c n) m, Plotable (VectorPlot v n) b, TypeableFloat n) => [v n] -> [(n, n)] -> ArrowOpts n -> m ()