Plots.Types.HeatMap

Documentation

data HeatMatrix

2D Array of Doubles.

Constructors

HeatMatrix

Fields

hmSize :: !(V2 Int)

hmFun :: V2 Int -> Double

mkHeatMatrix :: V2 Int -> (V2 Int -> Double) -> HeatMatrix

Create a heat matrix from an extent and a function to extract the values.

mkHeatMatrix' :: Foldable f => f (f Double) -> HeatMatrix

Create a heat matrix from a list of lists of values. All sublists should have the same length

normaliseHeatMatrix :: HeatMatrix -> ((Double, Double), HeatMatrix)

Normalise a heat matrix so all values lie in the (0,1) range. Also returns the lower and upper limit of the original heat matrix.

hmPoints :: IndexedTraversal' (V2 Int) HeatMatrix Double

Indexed traversal over the values of a HeatMatrix.

pixelHeatRender :: (Renderable (DImage n Embedded) b, TypeableFloat n) => HeatMatrix -> ColourMap -> QDiagram b V2 n Any

Render an heatmap as an image.

heatImage :: HeatMatrix -> ColourMap -> Image PixelRGB8

colourToPixel :: Colour Double -> PixelRGB8

pathHeatRender :: (Renderable (Path V2 n) b, TypeableFloat n) => HeatMatrix -> ColourMap -> QDiagram b V2 n Any

Render the heat map as squares.

data HeatMap b n

Constructors

HeatMap

Fields

hMatrix :: HeatMatrix

hStart :: P2 n

hSize :: V2 n

hGridSty :: Style V2 n

hGridVisible :: Bool

hLimits :: Maybe (Double, Double)

hDraw :: HeatMatrix -> ColourMap -> QDiagram b V2 n Any

class HasHeatMap f a b | a -> b where

Class of things that let you change the heatmap options.

Minimal complete definition

heatMapOptions

Methods

heatMapOptions :: LensLike' f a (HeatMap b (N a))

Lens onto the heatmap options.

heatMapGridVisible :: Functor f => LensLike' f a Bool

Whether there should be grid lines draw for the heat map. Default is False.

heatMapGridStyle :: Functor f => LensLike' f a (Style V2 (N a))

The style applied to the grid lines for the heat map, if they're visible.

heatMapSize :: Functor f => LensLike' f a (V2 (N a))

The size of each individual square in the heat map.

heatMapStart :: Functor f => LensLike' f a (P2 (N a))

The starting point for the heat map.

heatMapLimits :: Functor f => LensLike' f a (Maybe (Double, Double))

Limits (a,b) used on the data such that a is the start of the ColourMap and b is the end of the ColourMap. Default is (0,1).

heatMapRender :: Functor f => LensLike' f a (HeatMatrix -> ColourMap -> QDiagram b V2 (N a) Any)

Funtion used to render the heat map.

mkHeatMap :: (Renderable (Path V2 n) b, TypeableFloat n) => HeatMatrix -> HeatMap b n

Construct a Heatmap using the given HeatMatrix.

heatMap

Arguments

:: (VectorLike V2 Int i, TypeableFloat n, Typeable b, MonadState (Axis b V2 n) m, Renderable (Path V2 n) b)

=> i
extent of array

-> (i -> Double)
heat from index

-> State (Plot (HeatMap b n) b) ()
changes to plot options

-> m ()
add plot to Axis

Add a HeatMap plot using the extent of the heatmap and a generating function.

heatMap :: V2 Int     -> (V2 Int -> Double)     -> State (Plot (HeatMap b n)) () -> State (Axis b V2 n) ()
heatMap :: (Int, Int) -> ((Int, Int) -> Double) -> State (Plot (HeatMap b n)) () -> State (Axis b V2 n) ()