Plots.Types.Line

Trail plot

mkTrail :: (PointLike v n p, OrderedField n, Foldable f) => f p -> Located (Trail v n)

Construct a localed trail from a list of folable of points.

mkTrailOf :: (PointLike v n p, OrderedField n) => Fold s p -> s -> Located (Trail v n)

Construct a localed trail from a fold over points.

Path plot

mkPath :: (PointLike v n p, OrderedField n, Foldable f, Foldable g) => g (f p) -> Path v n

Construct a localed trail from a fold over points.

mkPathOf :: (PointLike v n p, OrderedField n) => Fold s t -> Fold t p -> s -> Path v n

Construct a localed trail from a fold over points.

GLinePlot plot

data GLinePlot v n a

mkGLinePlotOf :: (PointLike v n p, Fractional n) => Fold s a -> s -> (a -> p) -> GLinePlot v n a

Plot a general line plot give a fold

mkGLinePlot :: (PointLike v n p, Foldable f, Fractional n) => f a -> (a -> p) -> GLinePlot v n a

Plot a general line plot.

Line plot

type LinePlot v n = GLinePlot v n (Point v n)

mkLinePlotOf :: (PointLike v n p, Num n) => Fold s p -> s -> LinePlot v n

Make a line plot using the given fold.

mkLinePlot :: (PointLike v n p, Foldable f, Num n) => f p -> LinePlot v n

Make a line plot.

Helper functions

createStepData :: [(a, a)] -> [(a, a)]

Create a data for step plots.

Lenses

dotsonPoint :: HasPath a v n d => Lens' a Bool

pathStyle :: HasPath a v n d => Lens' a (Maybe (d -> Style V2 n))

Line Plot

linePlot :: (v ~ BaseSpace c, PointLike v n p, MonadState (Axis b c n) m, Plotable (LinePlot v n) b, Foldable f) => f p -> State (Plot (LinePlot v n) b) () -> m ()

linePlot' :: (v ~ BaseSpace c, PointLike v n p, MonadState (Axis b c n) m, Plotable (LinePlot v n) b, Foldable f) => f p -> m ()

Make a LinePlot and take a State on the plot to alter it's options

linePlotOf :: (v ~ BaseSpace c, PointLike v n p, MonadState (Axis b c n) m, Plotable (LinePlot v n) b) => Fold s p -> s -> State (Plot (LinePlot v n) b) () -> m ()

Add a LinePlot with the given name for the legend entry.

  myaxis = r2Axis &~ do
    linePlotL "blue team" pointData1
    linePlotL "red team" pointData2

linePlotL :: (v ~ BaseSpace c, PointLike v n p, MonadState (Axis b c n) m, Plotable (LinePlot v n) b, F.Foldable f) => String -> f p -> m () linePlotL l d = addPlotableL l (mkLinePlot d)

mkTrail version of line plot

linePlotOf' :: (v ~ BaseSpace c, PointLike v n p, MonadState (Axis b c n) m, Plotable (LinePlot v n) b) => Fold s p -> s -> m ()

Step plot

stepPlot :: (RealFloat n, Typeable n, Typeable b, Renderable (Path V2 n) b, MonadState (Axis b c n) m, BaseSpace c ~ V2) => [(n, n)] -> State (Plot (LinePlot V2 n) b) () -> m ()

stepPlot' :: (RealFloat n, Typeable n, Typeable b, Renderable (Path V2 n) b, MonadState (Axis b c n) m, BaseSpace c ~ V2) => [(n, n)] -> m ()

Gline plot

glinePlot :: (v ~ BaseSpace c, PointLike v n p, MonadState (Axis b c n) m, Plotable (GLinePlot v n a) b, Foldable f) => f a -> (a -> p) -> State (Plot (GLinePlot v n a) b) () -> m ()

glinePlot' :: (v ~ BaseSpace c, PointLike v n p, MonadState (Axis b c n) m, Plotable (GLinePlot v n a) b, Foldable f) => f a -> (a -> p) -> m ()