Plots.Util

Documentation

pathFromVertices :: (Metric v, OrderedField n) => [Point v n] -> Path v n

Type specialised version of fromVertices.

minMaxOf :: (Fractional a, Ord a) => Getting (Endo (Endo (V2 a))) s a -> s -> (a, a)

Minmax of a getter in the form V2 min max. Returns (V2 (-Infinity) Infinity) for empty folds.

enumFromToN :: Fractional n => n -> n -> Int -> [n]

enumFromToN a b n calculates a list from a to b in n steps.

whenever :: Bool -> (a -> a) -> a -> a

Apply a function if the predicate is true.

State helpers

(&=) :: MonadState s m => ASetter' s b -> State b a -> m () infix 3

Similar to '(%=)' but takes a state modification instead of a function.

(&~~) :: Monad m => s -> StateT s m a -> m s infix 1

Similar to '(&~)' but works with StateT and returns it in m.