Plots.Utils

Documentation

liftRecommend :: (a -> a -> a) -> Recommend a -> Recommend a -> Recommend a

Apply a function over two recommends or two commits. If only one of the values is a commit, the commit is used without applying the function.

fromCommit :: a -> Recommend a -> a

Extract a commit value, defaulting to the provided value when it is a recommend.

pathFromVertices :: (Metric v, OrderedField n) => [Point v n] -> Path v n

Type specialised version of fromVertices.

minMaxOf :: (Fractional a, Ord a) => Getting (Endo (Endo (V2 a))) s a -> s -> (a, a)

Minmax of a getter in the form V2 min max. Returns (V2 (-Infinity) Infinity) for empty folds.

enumFromToN :: Fractional n => n -> n -> Int -> [n]

enumFromToN a b n calculates a list from a to b in n steps.

State helpers

(&=) :: MonadState s m => ASetter' s b -> State b a -> m () infix 3

Similar to '(%=)' but takes a state modification instead of a function.

(&~~) :: Monad m => s -> StateT s m a -> m s infix 1

Similar to '(&~)' but works with StateT and returns it in m.