Plots.Axis.ColourBar

Description

Options for rendering a colour bar, either attached to an axis or rendered separately.

To change the colour map used for the colour bar see axisColourMap from Plots.Style.

Documentation

data ColourBar b n

Options for drawing a colour bar. Note that for an axis, the ColourMap is stored in the AxisStyle. These options are for other aspects of the bar, not the colours used.

Constructors

ColourBar

Fields

cbPlacement :: Placement

cbVisible :: Bool

cbTicks :: MajorTicks V2 n

cbGridLines :: MajorGridLines V2 n

cbTickLabels :: TickLabels b V2 n

cbDraw :: ColourMap -> QDiagram b V2 n Any

cbWidth :: n

cbLengthFun :: n -> n

cbGap :: n

cbStyle :: Style V2 n

defColourBar :: (Renderable (Text n) b, Renderable (Path V2 n) b, TypeableFloat n, Enum n) => ColourBar b n

The default colour bar.

class HasColourBar a b | a -> b where

Minimal complete definition

colourBar

Methods

colourBar :: Lens' a (ColourBar b (N a))

Lens onto the ColourBar.

colourBarDraw :: Lens' a (ColourMap -> QDiagram b V2 (N a) Any)

How to draw the colour bar. Should return a 10x1 box.

colourBarWidth :: Lens' a (N a)

The width (orthogonal to the colour bar direction) of the colour bar.

Default is 20.

colourBarLengthFunction :: Lens' a (N a -> N a)

Set the length of the colour bar given the length of the axis the colour bar is aligned to.

Default is id.

colourBarGap :: Lens' a (N a)

Gap between the axis and the colour bar (if rendered with an axis).

Default is 20.

colourBarStyle :: Lens' a (Style V2 (N a))

Style used for the outline of a colour bar.

addColourBar

Arguments

:: (TypeableFloat n, Renderable (Path V2 n) b)

=> BoundingBox V2 n
bounding box to place against

-> ColourBar b n

-> ColourMap

-> (n, n)

-> QDiagram b V2 n Any

Add a colour bar to an object, using the bounding box for the object.

renderColourBar

Arguments

:: (TypeableFloat n, Renderable (Path V2 n) b)

=> ColourBar b n
options for colour bar

-> ColourMap
map to use

-> (n, n)
bounds of the values on the colour bar

-> n
length of the colour bar

-> QDiagram b V2 n Any

Render a colour bar by it's self at a given width. Note this ignores colourBarGap and colourBarLengthFunction.

gradientColourBar :: (TypeableFloat n, Renderable (Path V2 n) b) => ColourMap -> QDiagram b V2 n Any

The colour bar generated by a gradient texture. This may not be supported by all backends.

pathColourBar :: (TypeableFloat n, Renderable (Path V2 n) b) => Int -> ColourMap -> QDiagram b V2 n Any

Construct a colour bar made up on n solid paths. The final diagram is 1 by 1, centred at in the center of the start of the bar.

flipX_Y :: Num n => V2 n -> V2 n

_reflectionX_Y :: (Additive v, R2 v, Num n) => Transformation v n

_reflectX_Y :: (InSpace v n t, R2 v, Transformable t) => t -> t