Plots.Axis.Grid

Description

Lines that go along the axis. Supports major and minor gird lines separately for each axis.

Grid lines

data GridLines v n

Type holding infomation about both major and minor grid lines.

class (HasMinorGridLines f a, HasMajorGridLines f a) => HasGridLines f a where

Minimal complete definition

gridLines

Methods

gridLines :: LensLike' f a (GridLines (V a) (N a))

data MajorGridLines v n

class HasMajorGridLines f a where

Minimal complete definition

majorGridLines

Methods

majorGridLines :: LensLike' f a (MajorGridLines (V a) (N a))

The options for how to draw the grid lines. This can be used on various levels of the axis:

majorGridLines :: Traversal' (Axis b c n)       (GridLines (BaseSpace c) n)
majorGridLines :: Lens'      (SingleAxis b v n) (GridLines v n)
majorGridLines :: Lens'      (GridLines v n)    (GridLines v n)

majorGridLinesFunction :: Functor f => LensLike' f a (GridLineFunction (N a))

The function to calculate location of the major grid lines given location of the major ticks and bounds.

majorGridLinesStyle :: Functor f => LensLike' f a (Style (V a) (N a))

The style applied to the major grid lines.

data MinorGridLines v n

class HasMinorGridLines f a where

Minimal complete definition

minorGridLines

Methods

minorGridLines :: LensLike' f a (MinorGridLines (V a) (N a))

The options for how to draw the grid lines. This can be used on various levels of the axis:

minorGridLines :: Traversal' (Axis b c n)       (GridLines (BaseSpace c) n)
minorGridLines :: Lens'      (SingleAxis b v n) (GridLines v n)
minorGridLines :: Lens'      (GridLines v n)    (GridLines v n)

minorGridLinesFunction :: Functor f => LensLike' f a (GridLineFunction (N a))

The function to calculate location of the minor grid lines given location of the minor ticks and bounds.

minorGridLinesStyle :: Functor f => LensLike' f a (Style (V a) (N a))

The style applied to the minor grid lines.

Extra traversals

gridLinesStyle :: (HasGridLines f a, Applicative f) => LensLike' f a (Style (V a) (N a))

Traversal over both the major and minor grid styles. This can be used at seversal levels in the Axis:

gridLinesVisible :: (HasGridLines f a, Applicative f) => LensLike' f a Bool

Traversal over both the major and minor grid styles.

gridLinesVisible :: Traversal' (Axis b c n) Bool
gridLinesVisible :: Traversal' (SingleAxis b v n) Bool
gridLinesVisible :: Traversal' (GridLines v n) Bool

hideGridLines :: (HasGridLines Identity a, MonadState a m) => m ()

Hide both major and minor grid lines.

hideGridLines :: Axis b c n       -> Axis b c n
hideGridLines :: SingleAxis b c n -> SingleAxis b c n
hideGridLines :: GridLines b c n  -> GirdLines b c n

showGridLines :: (HasGridLines Identity a, MonadState a m) => m ()

Show both major and minor grid lines.

showGridLines :: Axis b c n       -> Axis b c n
showGridLines :: SingleAxis b c n -> SingleAxis b c n
showGridLines :: GridLines b c n  -> GirdLines b c n

Grid line functions

type GridLineFunction n = [n] -> (n, n) -> [n]

A grid line function takes the positions of the respective ticks (minor ticks for minor grid lines, major ticks for major grid lines) and the bounds of the axis and returns the positions of the grid lines.

These functions are used in conjuction with majorGridLineFunction and minorGridLineFunction to control how the lines are drawn.

onTicksGridLineFunction :: GridLineFunction n

Place grid lines at the same position as the respective ticks. This is the Default.

emptyGridLineFunction :: GridLineFunction n

The GridLineFunction such that no grid lines appear.

See hideGridLines, majorGridLineVisible or minorGridLineVisible if you just want to hide the grid lines.