Plots.Axis.Labels

Description

Low level module defining types for axis labels and tick labels.

Axis line labels

class HasAxisLabel f a b | a -> b where

Minimal complete definition

axisLabel

Methods

axisLabel :: LensLike' f a (AxisLabel b (V a) (N a))

The options for the label of the axis. This can be used on various levels of the axis:

axisLabel :: Traversal' (Axis b c n)       (AxisLabel (BaseSpace c) n)
axisLabel :: Lens'      (SingleAxis b v n) (AxisLabel v n)
axisLabel :: Lens'      (AxisLabel v n)    (AxisLabel v n)

axisLabelText :: Functor f => LensLike' f a String

The text to use when labeling the axis.

axisLabelTextFunction :: Functor f => LensLike' f a (TextFunction b (V a) (N a))

The TextFunction to render the text of the axis label.

axisLabelGap :: Functor f => LensLike' f a (N a)

The gap between the axis and the labels, in the direction corresponding to the axisLabelPosition.

axisLabelStyle :: Functor f => LensLike' f a (Style (V a) (N a))

The Style to use on the rendered text.

axisLabelPosition :: Functor f => LensLike' f a AxisLabelPosition

The position the label will be placed parallel the axis.

axisLabelPlacement :: Functor f => LensLike' f a AxisLabelPosition

Whether the axis label should be placed inside or outside the axis.

axisLabelVisible :: Functor f => LensLike' f a Bool

Whether the axis label should be visible.

data AxisLabel b v n

data AxisLabelPosition

The position of the AxisLabel along the axis.

Constructors

MiddleAxisLabel

LowerAxisLabel

UpperAxisLabel

data AxisLabelPlacement

Whether the AxisLabel should be inside or ouside the axis.

Constructors

InsideAxisLabel

OutsideAxisLabel

Axis tick labels

data TickLabels b v n

class HasTickLabels f a b | a -> b where

Minimal complete definition

tickLabel

Methods

tickLabel :: LensLike' f a (TickLabels b (V a) (N a))

The options for the label of ticks. This can be used on various levels of the axis:

tickLabel :: Traversal' (Tick b c n)       (TickLabel (BaseSpace c) n)
tickLabel :: Lens'      (SingleAxis b v n) (TickLabel v n)
tickLabel :: Lens'      (TickLabel v n)    (TickLabel v n)

tickLabelTextFunction :: Functor f => LensLike' f a (TextFunction b (V a) (N a))

The TextFunction to render the text.

Default is atMajorTicks floatShow

tickLabelFunction :: Functor f => LensLike' f a (TickLabelFunction (N a))

The TextFunction to render the text.

Default is mkText.

tickLabelStyle :: Functor f => LensLike' f a (Style (V a) (N a))

The Style to use on the rendered text.

Default is fontSize (output 11).

tickLabelGap :: Functor f => LensLike' f a (N a)

The gap between the axis and the tick labels.

Default is 12.

tickLabelVisible :: Functor f => LensLike' f a Bool

Whether the axis label should be visible.

Default is True.

tickLabelPositions :: (HasTickLabels f a b, Settable f) => LensLike' f a [(N a, String)]

Setter over the final positions the major ticks. This is not as general as minorTicksFunction because you don't have access to the bounds but it can be useful when you know exactly what ticks you want to add or modify existing tick positions.

atMajorTicks :: (n -> String) -> TickLabelFunction n

Make a TickLabelFunction by specifying how to draw a single label from a position on the axis.