Plots.Axis.Line

Description

The lines that make up an axis.

Grid lines

data AxisLine v n

Information about position and style of axis lines.

class HasAxisLine f a where

Class of object that have an AxisLine.

Minimal complete definition

axisLine

Methods

axisLine :: LensLike' f a (AxisLine (V a) (N a))

Lens onto the AxisLine.

axisLineType :: Functor f => LensLike' f a AxisLineType

The position of the axis line around the axis.

Default is BoxAxisLine.

axisLineArrowOpts :: Functor f => LensLike' f a (Maybe (ArrowOpts (N a)))

The options for if you want the axis line to have arrows at the end.

Default is Nothing.

XXX (feature not currently implimented)

axisLineStyle :: Functor f => LensLike' f a (Style (V a) (N a))

The Style applied to the axis line

Axis line types

data AxisLineType

Where axis line for coordinate should be drawn. The Default is BoxAxisLine.

Constructors

BoxAxisLine

LeftAxisLine

MiddleAxisLine

RightAxisLine

NoAxisLine