Plots.Axis.Scale

Description

Determine how to scale an axis.

Axis scale

data AxisScaling n

Data type used that concerns everything to do with the size or scale of the axis.

data ScaleMode

How the axis should be scaled when not all dimensions are set.

Constructors

AutoScale

NoScale

Stretch

UniformScale UniformScaleStrategy

data UniformScaleStrategy

?

Constructors

AutoUniformScale

UnitOnly

ChangeVerticalLimits

ChangeHorizontalLimits

data Extending n

How much to extend the bounds beyond any inferred bounds.

Constructors

AbsoluteExtend n

RelativeExtend n

noExtend :: Num n => Extending n

Do not extend the axis beyond the inferred bounds.

class HasAxisScaling f a where

Class of things that have an AxisScaling.

Minimal complete definition

axisScaling

Methods

axisScaling :: LensLike' f a (AxisScaling (N a))

The way to scale in one direction.

scaleAspectRatio :: Functor f => LensLike' f a (Maybe (N a))

The ratio relative to other axis. If no ratios are set, the ratio is not enforced. If at least one is set, Nothing ratios are 1.

scaleMode :: Functor f => LensLike' f a ScaleMode

The mode to determine how to scale the bounds in a direction. Choose between AutoScale, NoScale, Stretch or UniformScale.

Default is AutoScale.

logScale :: Functor f => LensLike' f a LogScale

Whether the axis uses LogAxis or LinearAxis.

Default is LinearAxis.

axisExtend :: Functor f => LensLike' f a (Extending (N a))

How much to extend the bounds over infered bounds. This is ignored if a boundMax or boundMin is set.

boundMin :: Functor f => LensLike' f a (Maybe (N a))

The maximum bound the axis. There are helper functions for setting a minimum bound for a specific axis.

xMin :: Lens' (Axis b V2 Double) (Maybe Double)
yMin :: Lens' (Axis b V2 Double) (Maybe Double)

Default is Nothing.

boundMax :: Functor f => LensLike' f a (Maybe (N a))

The maximum bound the axis. There are helper functions for setting a maximum bound specific axis.

xMax :: Lens' (Axis b V2 Double) (Maybe Double)
yMax :: Lens' (Axis b V2 Double) (Maybe Double)
rMax :: Lens' (Axis b 'Polar Double) (Maybe Double)

Default is Nothing.

renderSize :: Functor f => LensLike' f a (Maybe (N a))

The size of the rendered axis. Default is Just 400.

Log scales

data LogScale

Should the axis be on a logarithmic scale. The Default is LinearAxis.

Constructors

LinearAxis

LogAxis

logNumber :: Floating a => LogScale -> a -> a

Log the number for LogAxis, do nothing for LinearAxis.

logPoint :: (Additive v, Floating n) => v LogScale -> Point v n -> Point v n

Transform a point according to the axis scale. Does nothing for linear scales.

logDeform :: (InSpace v n a, Foldable v, Floating n, Deformable a a) => v LogScale -> a -> a

Deform an object according to the axis scale. Does nothing for linear scales.

Low level calculations

These functions are used by Plots.Axis.Render.

calculateBounds

Arguments

:: OrderedField n

=> AxisScaling n
Scaling to use for this axis

-> Maybe (n, n)
Inferred bounds (from any plots)

-> (n, n)
Lower and upper bounds to use for this axis

Calculating the bounds for an axis.

calculateScaling

Arguments

:: (HasLinearMap v, OrderedField n, Applicative v)

=> v (AxisScaling n)
axis scaling options

-> BoundingBox v n
bounding box from the axis plots

-> (v (n, n), Transformation v n, Transformation v n)

Calculate the scaling for the axis.

The result returns:

The final bounds for the axis

scale to match desired scaleAspectRatio

scale to match desired asSizeSpec