Plots.Axis.Ticks

Description

Ticks for being placed on an axis or a ColourBar.

Major ticks

data MajorTicks v n

The big ticks on the axis line.

class HasMajorTicks f a where

Class of things that have a single MajorTicks.

Minimal complete definition

majorTicks

Methods

majorTicks :: LensLike' f a (MajorTicks (V a) (N a))

Lens onto the MajorTicks of something.

majorTicksFunction :: Functor f => LensLike' f a ((N a, N a) -> [N a])

The function used to place ticks for this axis, given the bounds of the axis. The result of these major ticks are also used as guides for MinorTicks, MajorGridLines and MinorGridLines.

Default is linearMinorTicks 5.

majorTicksAlignment :: Functor f => LensLike' f a TicksAlignment

Alignment of the major ticks. Choose between autoTicks (default), centreTicks, insideTicks or outsideTicks.

majorTicksLength :: Functor f => LensLike' f a (N a)

The total length the major ticks.

Default is 7.

majorTicksStyle :: Functor f => LensLike' f a (Style (V a) (N a))

The style used to render the major ticks.

Default is lwO 0.6 mempty (subject to change).

majorTicksHelper

Arguments

:: (Enum n, RealFrac n, Floating n)

=> [n]
Allowed numbers (up to powers of 10)

-> n
desired number of ticks

-> (n, n)
bounds

-> [n]
tick positions

Choose ticks whose step size is a multiple of 10 of the allowed numbers and tries to match the number of desired ticks.

logMajorTicks :: (Enum n, RealFrac n, Floating n) => n -> (n, n) -> [n]

Place n ticks at powers of 10 on the axis.

Minor ticks

data MinorTicks v n

The small ticks on the axis line.

class HasMinorTicks f a where

Class of things that have a single MinorTicks.

Minimal complete definition

minorTicks

Methods

minorTicks :: LensLike' f a (MinorTicks (V a) (N a))

Lens onto the MinorTicks of something.

minorTicksFunction :: Functor f => LensLike' f a ([N a] -> (N a, N a) -> [N a])

The function used to place ticks for this axis, given the result of majorTicksFunction and the bounds of the axis.

Default is linearMinorTicks 3.

minorTicksAlignment :: Functor f => LensLike' f a TicksAlignment

Alignment of the minor ticks. Choose between autoTicks (default), centreTicks, insideTicks or outsideTicks.

minorTicksLength :: Functor f => LensLike' f a (N a)

The total length the minor ticks.

Default is 3.

minorTicksStyle :: Functor f => LensLike' f a (Style (V a) (N a))

The style used to render the minor ticks.

Default is lwO 0.4 mempty (subject to change).

minorTicksHelper

Arguments

:: Fractional n

=> Int
Number of minor ticks between each major tick

-> [n]
Positions of major ticks

-> (n, n)
Bounds

-> [n]
Minor tick positions

Place n linear spaced ticks between each major tick.

Both major and minor ticks

data Ticks v n

Both MajorTicks and MinorTicks together.

class (HasMinorTicks f a, HasMajorTicks f a) => HasTicks f a where

Class of things with both MajorTicks and MinorTicks.

Minimal complete definition

bothTicks

Methods

bothTicks :: LensLike' f a (Ticks (V a) (N a))

ticksAlign :: (HasTicks f a, Applicative f) => LensLike' f a TicksAlignment

Traversal over both major and minor tick alignment.

ticksStyle :: (HasTicks f a, Applicative f) => LensLike' f a (Style (V a) (N a))

Traversal over both major and minor tick styles.

ticksVisible :: (HasTicks f a, Applicative f) => LensLike' f a Bool

Traversal over the visibility of both major and minor ticks.

Tick alignment

data TicksAlignment

Set the portion of the tick above and below the axis.

Constructors

TickSpec !Rational !Rational

AutoTick

autoTicks :: TicksAlignment

Set the tick type depending on the axis line position. centreTick for middleAxis, insideTick for everything else.

centreTicks :: TicksAlignment

Set the tick to be in the centre of the axis with total length of the corresponding tick length.

centerTicks :: TicksAlignment

Synonym for centreTicks.

insideTicks :: TicksAlignment

Align the ticks to be inside a box axis.

outsideTicks :: TicksAlignment

Align the ticks to be outside a box axis.

Helper functions

hideTicks :: HasTicks Identity a => a -> a

Hides the Minor ticks when trying to render something. This can be used on multiple types:

hideTicks :: Axis b c n       -> Axis b c n
hideTicks :: SingleAxis b v n -> SingleAxis b v n
hideTicks :: Ticks v n        -> Ticks v n
hideTicks :: MinorTicks v n   -> MinorTicks v n

majorTickPositions :: (HasMajorTicks f a, Settable f) => LensLike' f a [N a]

Setter over the final positions the major ticks. This is not as general as majorTicksFunction because you don't have access to the bounds but it can be useful when you know exactly what ticks you want to add or modify existing tick positions.

minorTickPositions :: (HasMinorTicks f a, Settable f) => LensLike' f a [N a]

Setter over the final positions the major ticks. This is not as general as minorTicksFunction because you don't have access to the bounds but it can be useful when you know exactly what ticks you want to add or modify existing tick positions.

linearMajorTicks :: (Enum n, RealFrac n, Floating n) => n -> (n, n) -> [n]

Ticks whose value ends in 1, 0.5, 0.25, 0.2 (*10^n).

linearMinorTicks :: (Enum n, Fractional n, Ord n) => n -> [n] -> (n, n) -> [n]

Position n minor ticks between each major tick.