Plots.Legend

Legend

data Legend b n

The data type to describe how to draw a legend. For legend entries see LegendEntry.

class HasLegend a b | a -> b where

Minimal complete definition

legend

Methods

legend :: Lens' a (Legend b (N a))

Lens onto the Legend of something.

legendPlacement :: Lens' a Placement

The Placement of the legend relative to the Axis.

legendGap :: Lens' a (N a)

The gap between the legend and the axis.

legendStyle :: Lens' a (Style V2 (N a))

The style applied to the surronding box of the legend.

legendSpacing :: Lens' a (N a)

The spacing between entries in the legend.

legendTextWidth :: Lens' a (N a)

The space given for the text in the legend.

legendTextFunction :: Lens' a (String -> QDiagram b V2 (N a) Any)

The function to generate the legend text.

legendTextStyle :: Lens' a (Style V2 (N a))

The style applied to the legend text.

legendOrientation :: Lens' a Orientation

The way the legend entries are listed. (This will likely be replaced by a grid-like system)

Drawing a legend

drawLegend

Arguments

:: (TypeableFloat n, Renderable (Path V2 n) b)

=> BoundingBox V2 n
bounding box to place legend against

-> [(QDiagram b V2 n Any, String)]
diagram pictures along with their key

-> Legend b n
options for drawing the legend

-> QDiagram b V2 n Any
rendered legend

Draw a legend to the bounding box using the legend entries and legend options.