Plots.Legend

Legend

data Legend b n

The data type to describe how to draw a legend. For legend entries see LegendEntry.

class HasLegend a b | a -> b where

Minimal complete definition

legend

Methods

legend :: Lens' a (Legend b (N a))

legendPosition :: Lens' a Position

The Position of the legend relative to the Axis.

legendAnchor :: Lens' a Anchor

The anchor for where the legend is placed.

legendGap :: Lens' a (V2 (N a))

The gap between the legend and the axis.

legendStyle :: Lens' a (Style V2 (N a))

The style applied to the surronding box of the legend.

legendSpacing :: Lens' a (N a)

The spacing between entries in the legend.

legendTextWidth :: Lens' a (N a)

The space given for the text in the legend.

legendTextFunction :: Lens' a (String -> QDiagram b V2 (N a) Any)

The function to generate the legend text.

legendTextStyle :: Lens' a (Style V2 (N a))

The style applied to the legend text.

legendOrientation :: Lens' a Orientation

The way the legend entries are listed. (This will likely be replaced by a grid-like system)

legendVisible :: Lens' a Bool

Whether the legend should be visible.

Legend configuration

Positioning

data Position

Constructors

North

NorthEast

East

SouthEast

South

SouthWest

West

NorthWest

Position Rational Rational
Position 0 0 = SouthWest, Position 1 1 = NorthWest

getPosition :: (InSpace V2 n a, Enveloped a, HasOrigin a, Fractional n) => Position -> a -> P2 n

Get the point from the Position on the bounding box of the enveloped object. Returns the origin if a has an empty envelope.

data Anchor

Constructors

AnchorTop

AnchorTopRight

AnchorRight

AnchorBottomRight

AnchorBottom

AnchorBottomLeft

AnchorLeft

AnchorTopLeft

Anchor Rational Rational
Anchor 0 0 = AnchorBottomLeft, mkP2 1 1 = AnchorTopRight

anchor :: (InSpace V2 n a, Alignable a, HasOrigin a, Floating n) => Anchor -> a -> a

Align an object using a given anchor.

alignTo :: (InSpace V2 n a, SameSpace a b, Enveloped a, HasOrigin a, Alignable b, HasOrigin b, Floating n) => Position -> a -> Anchor -> V2 n -> b -> b

A tool for aligned one object to another.

Drawing a legend

drawLegend :: (TypeableFloat n, Typeable b, Renderable (Path V2 n) b, Renderable (Text n) b) => BoundingBox V2 n -> [(QDiagram b V2 n Any, String)] -> Legend b n -> QDiagram b V2 n Any

Draw a legend to the bounding box using the legend entries and legend options.