Plots.Types

Description

This module defines the various types for holding plots:

PlotOptions b v n

Generic options all plots have.

PlotMods b v n

Includes PlotOptions along with modifications to the PlotStyle.

Plot p b

A rawPlot p grouped with a PlotMods.

DynamicPlot b v n

A wrapped up Plot so it can be stored in an Axis.

StyledPlot b v n

A DynamicPlot with a concrete PlotStyle, ready to be rendered.

As well as other things like the Plotable class, LegendEntries, HasOrientation and HasVisibility.

Plot options

data PlotOptions b v n

Data type for holding information all plots must contain.

class HasPlotOptions f a b | a -> b where

Class of things that have PlotOptions.

Minimal complete definition

plotOptions

Methods

plotOptions :: LensLike' f a (PlotOptions b (V a) (N a))

Lens onto the PlotOptions.

plotName :: Functor f => LensLike' f a Name

The Name applied to the plot. This gives a way to reference a specific plot in a rendered axis.

Default is mempty.

clipPlot :: Functor f => LensLike' f a Bool

Whether the plot should be clipped to the bounds of the axes.

Default is True.

legendEntries :: Functor f => LensLike' f a [LegendEntry b (V a) (N a)]

The legend entries to be used for the current plot.

Default is mempty.

plotTransform :: Functor f => LensLike' f a (Transformation (V a) (N a))

The transform applied to the plot once it's in the axis coordinates.

Default is mempty.

plotVisible :: Functor f => LensLike' f a Bool

Whether or not the plot should be shown. The BoundingBox of the plot will still affect the inferred axis bounds.

Default is True.

key :: (HasPlotOptions Identity a b, MonadState a m, Num (N a)) => String -> m ()

Add a LegendEntry to something with PlotOptions using the String as the legendText and a DefaultLegendPic. Here are some typical examples:

key :: String -> State (Plot (ScatterPlot v n) b) ()
key :: String -> State (DynamicPlot b v n) ()
key :: String -> State (PlotMods b v n) ()

If you only care about the name of the legend, use key.

addLegendEntry :: (HasPlotOptions Identity a b, MonadState a m) => LegendEntry b (V a) (N a) -> m ()

Add a LegendEntry to something with PlotOptions. Here are some typical examples:

addLegendEntry :: LegendEntry b v n -> State (Plot (ScatterPlot v n) b) ()
addLegendEntry :: LegendEntry b v n -> State (DynamicPlot b v n) ()

If you only care about the name of the legend, use key.

Plot modifications

data PlotMods b v n

A PlotOptions with modifications to a PlotStyle.

plotMods :: Lens' (Plot p b) (PlotMods b (V p) (N p))

The modifications to the PlotOptions and PlotStyle in a Plot.

Plotable class

class (Typeable p, Enveloped p) => Plotable p b where

Class defining how plots should be rendered.

Minimal complete definition

renderPlotable

Methods

renderPlotable :: InSpace v n p => AxisSpec v n -> PlotStyle b v n -> p -> QDiagram b v n Any

Render a plot according to the AxisSpec, using the PlotStyle.

defLegendPic :: InSpace v n p => PlotStyle b v n -> p -> QDiagram b v n Any

The default legend picture when the LegendPic is DefaultLegendPic.

Plot types

Parameterised plot

data Plot p b

A parameterised plot, together with a PlotMods. This type has an instance of many classes for modifying specific plots.

mkPlot :: (Additive (V p), Num (N p)) => p -> Plot p b

Make a Plot with Default PlotOptions.

rawPlot :: SameSpace p p' => Lens (Plot p b) (Plot p' b) p p'

Lens onto the raw Plotable inside a Plot.

Dynamic plot

data DynamicPlot b v n where

A wrapped up Plot, used to store plots in an Axis.

Constructors

DynamicPlot :: (InSpace v n p, Plotable p b) => Plot p b -> DynamicPlot b v n

_DynamicPlot :: (Plotable p b, Typeable b) => Prism' (DynamicPlot b (V p) (N p)) (Plot p b)

Prism for a DynamicPlot.

dynamicPlot :: forall p b. (Typeable p, Typeable b) => Traversal' (DynamicPlot b (V p) (N p)) (Plot p b)

Traversal over the dynamic plot without the Plotable constraint _DynamicPlot has.

dynamicPlotMods :: Lens' (DynamicPlot b v n) (PlotMods b v n)

The modifications to the PlotOptions and PlotStyle in a DynamicPlot.

Styled plot

data StyledPlot b v n

A DynamicPlot with a concrete style. This is suitable for being rendered with renderStyledPlot and get extract the legend entries with styledPlotLegend.

You can make a StyledPlot with styleDynamic

styledPlot :: forall p b. Typeable p => Traversal' (StyledPlot b (V p) (N p)) p

Traversal over a raw plot of a styled plot. The type of the plot must match for the traversal to be succesful.

styleDynamic :: PlotStyle b v n -> DynamicPlot b v n -> StyledPlot b v n

Give a DynamicPlot a concrete PlotStyle.

renderStyledPlot :: TypeableFloat n => AxisSpec V2 n -> StyledPlot b V2 n -> QDiagram b V2 n Any

Render a StyledPlot given an and AxisSpec.

singleStyledPlotLegend

Arguments

:: StyledPlot b v n

-> [(n, QDiagram b v n Any, String)]
(z-order, legend pic, legend text)

Get the legend rendered entries from a single styled plot. The resulting entries are in no particular order. See also styledPlotLegends.

styledPlotLegends

Arguments

:: Ord n

=> [StyledPlot b v n]

-> [(QDiagram b v n Any, String)]
[(legend pic, legend text)]

Render a list of legend entries, in order.

Miscellaneous

Visibility

class HasVisibility a where

Class of objects that can be hidden.

Minimal complete definition

visible

Methods

visible :: Lens' a Bool

Lens onto whether an object should be visible when rendered.

hidden :: Lens' a Bool

The opposite of visible.

hide :: (MonadState s m, HasVisibility a) => ASetter' s a -> m ()

Set visible to False for the given setter.

hide minorTicks          :: State (Axis b v n) ()
hide (xAxis . gridLines) :: State (Axis b v n) ()

display :: (MonadState s m, HasVisibility a) => ASetter' s a -> m ()

Set visible to True for the given setter.

display minorGridLines :: State (Axis b v n) ()
display colourBar      :: State (Axis b v n) ()

Orientation

data Orientation

Constructors

Horizontal

Vertical

class HasOrientation a where

Class of things that have an orientation.

Minimal complete definition

orientation

Methods

orientation :: Lens' a Orientation

Lens onto the orientation of an object.

orient :: HasOrientation o => o -> a -> a -> a

Pick the first a if the object has Horizontal orientation and the second a if the object has a Vertical orientation.

horizontal :: HasOrientation a => Lens' a Bool

Lens onto whether an object's orientation is horizontal.

vertical :: HasOrientation a => Lens' a Bool

Lens onto whether an object's orientation is vertical.

Legend entries

data LegendEntry b v n

Data type for holding a legend entry.

data LegendPic b v n

Type allowing use of the default legend picture (depending on the plot) or a custom legend picture with access to the PlotStyle.

Constructors

DefaultLegendPic

CustomLegendPic (PlotStyle b v n -> QDiagram b v n Any)

mkLegendEntry :: Num n => String -> LegendEntry b v n

Make a legend entry with a default legendPicture and legendPrecedence 0 using the string as the legendText.

legendPicture :: Lens' (LegendEntry b v n) (LegendPic b v n)

The picture used in the legend entry.

legendText :: Lens' (LegendEntry b v n) String

The text used in the legend entry.

legendPrecedence :: Lens' (LegendEntry b v n) n

The order in which the legend entries are rendered. If precedence are equal, they entries are put in the order they are added to the axis.

Default is 0.

Axis spec

data AxisSpec v n

Information from the Axis necessary to render a Plotable.

Constructors

AxisSpec

Fields

_specBounds :: v (n, n)

_specTrans :: Transformation v n

_specScale :: v LogScale

_specColourMap :: ColourMap

specTrans :: forall v n. Lens' (AxisSpec v n) (Transformation v n)

specBounds :: forall v n. Lens' (AxisSpec v n) (v (n, n))

specScale :: forall v n. Lens' (AxisSpec v n) (v LogScale)

scaleNum :: Floating n => (n, n) -> LogScale -> n -> n

Scale a number by log10-ing it and linearly scaling it so it's within the same range.

specPoint :: (Applicative v, Additive v, Floating n) => AxisSpec v n -> Point v n -> Point v n

Apply log scaling and the transform to a point.

specColourMap :: forall v n. Lens' (AxisSpec v n) ColourMap

Positioning

data Placement

A Position is a point on an axis together with an anchor and a direction for the gap.

Constructors

Placement

Fields

pAt :: V2 Rational

pAnchor :: V2 Rational

pGapDir :: Direction V2 Rational

class HasPlacement a where

Minimal complete definition

placement

Methods

placement :: Lens' a Placement

placementAt :: Lens' a (V2 Rational)

The position relative to the axis. V2 0 0 corresponds to the bottom left corner, V2 1 1 is the top right corner.

placementAnchor :: Lens' a (V2 Rational)

The anchor used for the object being positioned. V2 0 0 corresponds to the bottom left corner, V2 1 1 is the top right corner.

gapDirection :: Lens' a (Direction V2 Rational)

The direction to extend the gap when positioning.

class HasGap a where

Minimal complete definition

gap

Methods

gap :: Lens' a (N a)

The value of the gap when rendering.

placeAgainst :: (InSpace V2 n a, SameSpace a b, Enveloped a, HasOrigin b, Alignable b) => a -> Placement -> n -> b -> b

A tool for aligned one object to another. Positions b around the bounding box of a by translating b.

Common positions

Inside positions

topLeft :: Placement

top :: Placement

topRight :: Placement

left :: Placement

right :: Placement

bottomLeft :: Placement

bottom :: Placement

bottomRight :: Placement

Outside positions

leftAbove :: Placement

leftTop :: Placement

leftMid :: Placement

leftBottom :: Placement

leftBelow :: Placement

midAbove :: Placement

midBelow :: Placement

rightAbove :: Placement

rightTop :: Placement

rightMid :: Placement

rightBottom :: Placement

rightBelow :: Placement