Plots.Types.HeatMap

Description

Documentation

data HeatMap b n

A mapping from points in a 2D axis do Doubles. These Doubles are converted to colours using the axis ColourMap.

heatMap

Arguments

:: (VectorLike V2 Int i, TypeableFloat n, Typeable b, MonadState (Axis b V2 n) m, Renderable (Path V2 n) b)

=> i
extent of array

-> (i -> Double)
heat from index

-> State (Plot (HeatMap b n) b) ()
changes to plot options

-> m ()
add plot to Axis

Add a HeatMap plot using the extent of the heatmap and a generating function.

heatMap :: V2 Int     -> (V2 Int -> Double)     -> State (Plot (HeatMap b n)) () -> State (Axis b V2 n) ()
heatMap :: (Int, Int) -> ((Int, Int) -> Double) -> State (Plot (HeatMap b n)) () -> State (Axis b V2 n) ()

heatMap'

Arguments

:: (VectorLike V2 Int i, TypeableFloat n, Typeable b, MonadState (Axis b V2 n) m, Renderable (Path V2 n) b)

=> i
extent of array

-> (i -> Double)
heat from index

-> m ()
add plot to Axis

Add a HeatMap plot using the extent of the heatmap and a generating function without changes to the heap map options.

heatMap :: V2 Int     -> (V2 Int -> Double)     -> State (Axis b V2 n) ()
heatMap :: (Int, Int) -> ((Int, Int) -> Double) -> State (Axis b V2 n) ()

Lenses

class HasHeatMap f a b | a -> b where

Class of things that let you change the heatmap options.

Minimal complete definition

heatMapOptions

Methods

heatMapOptions :: LensLike' f a (HeatMap b (N a))

Lens onto the heatmap options.

heatMapGridVisible :: Functor f => LensLike' f a Bool

Whether there should be grid lines draw for the heat map.

Default is False.

heatMapGridStyle :: Functor f => LensLike' f a (Style V2 (N a))

The style applied to the grid lines for the heat map, if they're visible.

Default is mempty.

heatMapSize :: Functor f => LensLike' f a (V2 (N a))

The size of each individual square in the heat map.

Default is V2 1 1.

heatMapExtent :: (Functor f, Fractional (N a)) => LensLike' f a (V2 (N a))

The size of the full extend of the heat map.

Default is extent of the heat matrix.

heatMapStart :: Functor f => LensLike' f a (P2 (N a))

The starting point at the bottom left corner of the heat map.

Default is origin

heatMapCentre :: (Functor f, Fractional (N a)) => LensLike' f a (P2 (N a))

The center point of the heat map.

heatMapLimits :: Functor f => LensLike' f a (Maybe (Double, Double))

Limits (a,b) used on the data such that a is the start of the ColourMap and b is the end of the ColourMap. Default is (0,1).

heatMapRender :: Functor f => LensLike' f a (HeatMatrix -> ColourMap -> QDiagram b V2 (N a) Any)

Funtion used to render the heat map. See pathHeatRender and pixelHeatRender.

Default is pathHeatRender.

Heat matrix

data HeatMatrix

2D Array of Doubles.

heatImage :: Int -> HeatMatrix -> ColourMap -> Image PixelRGB8

hmPoints :: IndexedTraversal' (V2 Int) HeatMatrix Double

Indexed traversal over the values of a HeatMatrix.

hmSize :: HeatMatrix -> V2 Int

normaliseHeatMatrix :: HeatMatrix -> ((Double, Double), HeatMatrix)

Normalise a heat matrix so all values lie in the (0,1) range. Also returns the lower and upper limit of the original heat matrix.

Rendering functions

pixelHeatRender :: (Renderable (DImage n Embedded) b, TypeableFloat n) => Int -> HeatMatrix -> ColourMap -> QDiagram b V2 n Any

Render an heatmap as an ImageRGB8 with n pixels per heat matrix point.

pathHeatRender :: (Renderable (Path V2 n) b, TypeableFloat n) => HeatMatrix -> ColourMap -> QDiagram b V2 n Any

Render the heat map as squares.

Low level construction

mkHeatMap :: (Renderable (Path V2 n) b, TypeableFloat n) => HeatMatrix -> HeatMap b n

Construct a Heatmap using the given HeatMatrix.

mkHeatMatrix :: V2 Int -> (V2 Int -> Double) -> HeatMatrix

Create a heat matrix from an extent and a function to extract the values.

mkHeatMatrix' :: Foldable f => f (f Double) -> HeatMatrix

Create a heat matrix from a list of lists of values. All sublists should have the same length