Plots.Types.Line

Description

A line plot is simply a Path used as a plot. This module contains helpers adding path plots. For line plots with markers, see Scatter.

Documentation

trailPlot

Arguments

:: (BaseSpace c ~ v, Plotable (Path v n) b, MonadState (Axis b c n) m)

=> Trail v n
trail to plot

-> State (Plot (Path v n) b) ()
changes to plot options

-> m ()
add plot to the Axis

Add a Trail as a Plot to an Axis.

trailPlot'

Arguments

:: (BaseSpace c ~ v, Plotable (Path v n) b, MonadState (Axis b c n) m)

=> Trail v n
trail to plot

-> m ()
add plot to the Axis

Add a Trail as a Plot to an Axis without changes to the plot options.

pathPlot

Arguments

:: (BaseSpace c ~ v, Plotable (Path v n) b, MonadState (Axis b c n) m)

=> Path v n
path to plot

-> State (Plot (Path v n) b) ()
changes to plot options

-> m ()
add plot to the Axis

Add a Path as a Plot to an Axis.

pathPlot'

Arguments

:: (BaseSpace c ~ v, Plotable (Path v n) b, MonadState (Axis b c n) m)

=> Path v n
path to plot

-> m ()
add plot to the Axis

Add a Path as a Plot to an Axis without changes to the plot options.

Line plots from points

linePlot

Arguments

:: (BaseSpace c ~ v, Metric v, Foldable f, PointLike v n p, Plotable (Path v n) b, MonadState (Axis b c n) m)

=> f p
points to turn into trail

-> State (Plot (Path v n) b) ()
changes to plot options

-> m ()
add plot to the Axis

Add a Path plot from a list of points.

linePlot'

Arguments

:: (BaseSpace c ~ v, Metric v, Foldable f, PointLike v n p, Plotable (Path v n) b, MonadState (Axis b c n) m)

=> f p
points to turn into trail

-> m ()
add plot to the Axis

Add a Path plot from a list of points.

smoothLinePlot

Arguments

:: (BaseSpace c ~ v, Foldable f, Metric v, PointLike v n p, Plotable (Path v n) b, Fractional (v n), MonadState (Axis b c n) m)

=> f p
points to turn into trail

-> State (Plot (Path v n) b) ()
changes to plot options

-> m ()
add plot to the Axis

Add a smooth Path plot from a list of points using cubicSpline.

smoothLinePlot'

Arguments

:: (BaseSpace c ~ v, Foldable f, PointLike v n p, Plotable (Path v n) b, Fractional (v n), MonadState (Axis b c n) m)

=> f p
points to turn into trail

-> m ()
add plot to the Axis

Add a smooth Path plot from a list of points using cubicSpline without changes to the plot options.

Construction utilities

Trails

mkTrail :: (PointLike v n p, OrderedField n, Foldable f) => f p -> Located (Trail v n)

Construct a localed trail from a list of folable of points.

mkTrailOf :: (PointLike v n p, OrderedField n) => Fold s p -> s -> Located (Trail v n)

Construct a localed trail from a fold over points.

Paths

mkPath :: (PointLike v n p, OrderedField n, Foldable f, Foldable g) => g (f p) -> Path v n

Construct a localed trail from a fold over points.

mkPathOf :: (PointLike v n p, OrderedField n) => Fold s t -> Fold t p -> s -> Path v n

Construct a localed trail from a fold over points.