Plots.Types.Points

Polar scatter plot

data GPointsPlot n

mkPointsPlot :: (RealFloat n, PointLike V2 n (Polar n), Num n) => [(n, Angle n)] -> GPointsPlot n

Plot a polar scatter plot given a list of radius and angle.

Lenses

doFill :: HasPoints a n => Lens' a Bool

Points plot

pointsPlot :: (v ~ BaseSpace c, v ~ V2, PointLike v n (Polar n), MonadState (Axis b c n) m, Plotable (GPointsPlot n) b, RealFloat n) => [(n, Angle n)] -> State (Plot (GPointsPlot n) b) () -> m ()

pointsPlot' :: (v ~ BaseSpace c, v ~ V2, PointLike v n (Polar n), MonadState (Axis b c n) m, Plotable (GPointsPlot n) b, RealFloat n) => [(n, Angle n)] -> m ()

Make a PointsPlot and take a State on the plot to alter it's options

  myaxis = polarAxis &~ do
    pointsPlot' pointData1 $ do
      addLegendEntry "data 1"
      doFill .= True